requirements.txt
click
rich
ruamel.yaml
streamlit
pytest
